backend/requirements.txt
flask
flask-cors
python-dotenv
openai
pdfplumber
python-docx
reportlab
gunicorn
